pyproject.toml
[project]
name = "deepglm"
version = "0.1.0"
description = "DeepGLM Research Agent - A streamlined research and knowledge work assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.4",
    "tavily-python>=0.7.17",
    "python-dotenv>=1.0.0",
    "langchain-openai>=1.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]  # Ignore line too long errors for docstrings

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

